requirements.txt
Django
Pillow
gunicorn
whitenoise

django-heroku
dj-database-url
django-phone-field
